+++ R/textbox-grob.R
@@ -172,6 +172,8 @@ textbox_grob <- function(text, x = NULL, y = NULL,
 
   # make sure we can handle input text even if provided as factor
   text <- as.character(text)
+  # convert NAs to empty strings
+  text <- ifelse(is.na(text), "", text)
 
   # determine orientation and adjust accordingly
   orientation <- match.arg(orientation)